+++ src/observer/storage/clog/log_entry.cpp
@@ -72,7 +72,7 @@ RC LogEntry::init(LSN lsn, LogModule::Id module_id, vector<char> &&data)
 
 RC LogEntry::init(LSN lsn, LogModule module, vector<char> &&data)
 {
-  if (data.size() > max_payload_size()) {
+  if (static_cast<int32_t>(data.size()) > max_payload_size()) {
     LOG_DEBUG("log entry size is too large. size=%d, max_payload_size=%d", data.size(), max_payload_size());
     return RC::INVALID_ARGUMENT;
   }
